# Conflicts:
#	nekoyume/Assets/_Scripts/UI/Module/Button/ConditionalCostButton.cs
#	nekoyume/Assets/_Scripts/UI/Module/Button/SimpleCostButton.cs
#	nekoyume/Assets/_Scripts/UI/Module/CombinationSlot.cs
#	nekoyume/Assets/_Scripts/UI/Module/Currency/ActionPoint.cs
#	nekoyume/Assets/_Scripts/UI/Module/Currency/ApPotion.cs
#	nekoyume/Assets/_Scripts/UI/Module/Currency/Hourglass.cs
#	nekoyume/Assets/_Scripts/UI/Widget/Popup/MaterialNavigationPopup.cs
#	nekoyume/Assets/_Scripts/UI/Widget/ShopSell.cs

+++ nekoyume/Assets/_Scripts/UI/Module/CombinationSlot.cs
@@ -358,7 +358,7 @@ namespace Nekoyume.UI.Module
             }
 
             var inventory = States.Instance.CurrentAvatarState.inventory;
-            var count = inventory.GetUsableItemCount((int)CostType.Hourglass, blockIndex);
+            var count = inventory.GetUsableItemCount(CostType.Hourglass, blockIndex);
             hourglassCountText.text = cost.ToString();
             hourglassCountText.color = count >= cost
                 ? Palette.GetColor(ColorType.ButtonEnabled)
